# internal/handler/openapi_embed.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "RideMatch API",
    "description": "Backend API for RideMatch — a multi-modal (car, okada, keke, bus) ride/logistics matching platform for Nigeria. Authentication is phone number + OTP based; there is no password field. All responses use a single envelope shape: `{ success, message, data, error }`.",
    "version": "1.0.0",
    "contact": {
      "name": "Joshua Olumoye"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1",
      "description": "Local development"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access token returned from /auth/otp/verify or /auth/token/refresh. Send as `Authorization: Bearer <token>`."
      }
    },
    "schemas": {
      "APIResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {},
          "error": {
            "type": "string"
          }
        }
      },
      "RequestOTPRequest": {
        "type": "object",
        "required": [
          "phone"
        ],
        "properties": {
          "phone": {
            "type": "string",
            "example": "08012345678",
            "description": "Any common Nigerian phone format: 080..., 234..., or +234..."
          }
        }
      },
      "RequestOTPResponse": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "example": "+2348012345678"
          },
          "expires_in_seconds": {
            "type": "integer",
            "example": 300
          },
          "resend_in_seconds": {
            "type": "integer",
            "example": 60
          }
        }
      },
      "VerifyOTPRequest": {
        "type": "object",
        "required": [
          "phone",
          "code"
        ],
        "properties": {
          "phone": {
            "type": "string",
            "example": "08012345678"
          },
          "code": {
            "type": "string",
            "example": "048213",
            "minLength": 6,
            "maxLength": 6
          },
          "name": {
            "type": "string",
            "example": "Joshua Olumoye",
            "description": "Used only on first-time registration; ignored on login."
          }
        }
      },
      "AuthTokens": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string"
          },
          "refresh_token": {
            "type": "string"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time"
          },
          "token_type": {
            "type": "string",
            "example": "Bearer"
          }
        }
      },
      "UserResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "phone": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "photo_url": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "user",
              "driver",
              "admin"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "suspended",
              "banned"
            ]
          },
          "rating_average": {
            "type": "number",
            "format": "float"
          },
          "is_driver": {
            "type": "boolean"
          }
        }
      },
      "VerifyOTPResponse": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserResponse"
          },
          "tokens": {
            "$ref": "#/components/schemas/AuthTokens"
          },
          "is_new_user": {
            "type": "boolean"
          }
        }
      },
      "RefreshTokenRequest": {
        "type": "object",
        "required": [
          "refresh_token"
        ],
        "properties": {
          "refresh_token": {
            "type": "string"
          }
        }
      },
      "LogoutRequest": {
        "type": "object",
        "required": [
          "refresh_token"
        ],
        "properties": {
          "refresh_token": {
            "type": "string"
          }
        }
      },
...
      "DriverProfileResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "user_id": {
            "type": "string",
            "format": "uuid"
          },
          "vehicle_type": {
            "type": "string",
            "enum": [
              "car",
              "okada",
              "keke",
              "bus"
            ]
          },
          "plate_number": {
            "type": "string"
          },
          "vehicle_photo_url": {
            "type": "string"
          },
          "verification_status": {
            "type": "string",
            "enum": [
              "pending",
              "approved",
              "rejected"
            ]
          },
          "verification_note": {
            "type": "string"
          },
          "is_online": {
            "type": "boolean"
          },
          "has_active_subscription": {
            "type": "boolean"
          },
          "subscription_active_until": {
            "type": "string",
            "format": "date-time"
          },
          "rating_average": {
            "type": "number",
            "format": "float"
          }
        }
      },
...
      "AdminVerifyDriverRequest": {
        "type": "object",
        "required": [
          "approve"
        ],
        "properties": {
          "approve": {
            "type": "boolean"
          },
          "note": {
            "type": "string"
          }
        }
      },
      "AdminActivateSubscriptionRequest": {
        "type": "object",
        "required": [
          "days"
        ],
        "properties": {
          "days": {
            "type": "integer",
            "minimum": 1,
            "maximum": 365,
            "example": 1
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/auth/otp/request": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Request a login/registration OTP",
        "description": "Sends a 6-digit one-time code by SMS to the given Nigerian phone number. Used for both first-time registration and subsequent logins — the same endpoint covers both. Subject to a resend cooldown (default 60s) per phone number.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestOTPRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Code sent",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/APIResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/RequestOTPResponse"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid phone number format"
          },
          "429": {
            "description": "Resend cooldown still active"
          }
        }
      }
    },
    "/auth/otp/verify": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Verify an OTP and obtain access/refresh tokens",
        "description": "Verifies the 6-digit code sent to a phone number. If no account exists for the phone number yet, one is created automatically (pass `name` on first verification). Returns a JWT access token and an opaque refresh token.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyOTPRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/APIResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/VerifyOTPResponse"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Missing/invalid fields, incorrect or expired code"
          },
          "403": {
            "description": "Account suspended or banned"
          },
          "429": {
            "description": "Too many incorrect attempts"
          }
        }
      }
    },
    "/auth/token/refresh": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Exchange a refresh token for a new token pair",
        "description": "Rotates the refresh token: the submitted token is revoked and a new access + refresh token pair is issued. The old refresh token cannot be reused.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "New token pair",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/APIResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/AuthTokens"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Invalid, expired, or already-used refresh token"
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Log out (revoke a refresh token)",
        "description": "Revokes the given refresh token, ending that session. Other devices/sessions for the same user are unaffected.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogoutRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Logged out"
          }
        }
      }
    },
    "/users/me": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get the authenticated user's profile",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Profile",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/APIResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/UserResponse"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token"
          }
        }
      }
    },
...
    "/admin/drivers/{id}/verify": {
      "patch": {
        "tags": [
          "Admin"
        ],
        "summary": "Approve or reject a driver's documents",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminVerifyDriverRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Verification updated",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/APIResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/DriverProfileResponse"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Caller is not an admin"
          },
          "404": {
            "description": "Driver not found"
          }
        }
      }
    },
    "/admin/drivers/{id}/subscription": {
      "patch": {
        "tags": [
          "Admin"
        ],
        "summary": "Manually activate/extend a driver's subscription",
        "description": "Stand-in for the Flutterwave payment webhook (not yet wired up) — grants the driver N days of platform access from now, or extends from their current expiry if still active.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminActivateSubscriptionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Subscription activated",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/APIResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/DriverProfileResponse"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Caller is not an admin"
          },
          "404": {
            "description": "Driver not found"
          }
        }
      }
    },
...
  }
}